Consolidation of block lists corresponding to RGB into a single file
